pyproject.toml
[tool.poetry]
name = "tts-server"
version = "3.0.0"
description = ""
authors = ["Nesswit <rishubil@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.8,<3.13"
Flask = "^3.0.1"
requests = "^2.31.0"
g2pkk = "^0.1.2"
python-mecab-ko = "^1.3.3"
onnxruntime = ">=1.11.0,<2"
piper-tts = "^1.2.0"

[tool.poetry.group.dev.dependencies]
poethepoet = "^0.24.4"
ruff = "^0.1.14"

[tool.poe]
envfile = ".env"

[tool.poe.tasks]
server = "python -u data/src/main.py"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
